alexaLambda/src/vergissmeinicht/alexa/ResumePlaybackRequestHandler.java
package vergissmeinicht.alexa;

import com.amazon.ask.dispatcher.request.handler.HandlerInput;
import com.amazon.ask.dispatcher.request.handler.RequestHandler;
import com.amazon.ask.model.Response;

import java.util.Optional;

import static com.amazon.ask.request.Predicates.intentName;

public class ResumePlaybackRequestHandler implements RequestHandler {
    @Override
    public boolean canHandle(HandlerInput input) {
        return input.matches(intentName("AMAZON.ResumeIntent"));
    }

    @Override
    public Optional<Response> handle(HandlerInput input) {
        return input.getResponseBuilder()
                .withSpeech("Das Fortsetzen pausierter Streams ist zum jetzigen Zeitpunkt leider noch nicht möglich.")
                .build();
    }
}
